Behavify and robustify ajax message test.
