+++ kalite/distributed/tests/__init__.py
@@ -1,4 +1,3 @@
-from ajax_messages_test import *
 from code_tests import *
 from url_tests import *
 from browser_tests import *
